split packageTabs
