+++ packages/admin-ui/src/pages-new/ai/packages/tabs/info/index.ts
@@ -1,0 +1,1 @@
+export { InfoTab } from "./InfoTab";
