+++ packages/admin-ui/src/pages-new/ai/packages/tabs/network/index.ts
@@ -1,0 +1,1 @@
+export { NetworkTab } from "./NetworkTab";
